Move tableCells and columns to ExpandedCardContent Fix search select bug
